+++ src/main/kotlin/dev/aaronhowser/mods/aaron/menu/BaseScreen.kt
@@ -20,16 +20,22 @@ abstract class BaseScreen<M : AbstractContainerMenu>(
 	protected open val showInventoryLabel = true
 
 	protected val rightPos: Int
-		get() = this.leftPos + this.imageWidth
+		get() = leftPos + imageWidth
 	protected val bottomPos: Int
-		get() = this.topPos + this.imageHeight
+		get() = topPos + imageHeight
+
+	open val titleLabelOffsetX: Int = 0
+	open val titleLabelOffsetY: Int = 0
+
+	open val inventoryLabelOffsetX: Int = 0
+	open val inventoryLabelOffsetY: Int = 0
 
 	final override fun init() {
-		this.imageWidth = background.width
-		this.imageHeight = background.height
+		imageWidth = background.width
+		imageHeight = background.height
 
-		this.leftPos = (this.width - this.imageWidth) / 2
-		this.topPos = (this.height - this.imageHeight) / 2
+		leftPos = (width - imageWidth) / 2
+		topPos = (height - imageHeight) / 2
 
 		baseInit()
 	}
@@ -38,11 +44,25 @@ abstract class BaseScreen<M : AbstractContainerMenu>(
 
 	override fun renderLabels(guiGraphics: GuiGraphics, mouseX: Int, mouseY: Int) {
 		if (showTitleLabel) {
-			guiGraphics.drawString(this.font, this.title, this.titleLabelX, this.titleLabelY, 4210752, false)
+			guiGraphics.drawString(
+				font,
+				title,
+				titleLabelX + titleLabelOffsetX,
+				titleLabelY + titleLabelOffsetY,
+				4210752,
+				false
+			)
 		}
 
 		if (showInventoryLabel) {
-			guiGraphics.drawString(this.font, this.playerInventoryTitle, this.inventoryLabelX, this.inventoryLabelY, 4210752, false)
+			guiGraphics.drawString(
+				font,
+				playerInventoryTitle,
+				inventoryLabelX + inventoryLabelOffsetX,
+				inventoryLabelY + inventoryLabelOffsetY,
+				4210752,
+				false
+			)
 		}
 	}